requirements.txt
Flask>=3.0,<4
google-api-python-client>=2.0,<3
google-auth>=2.0,<3
google-auth-oauthlib>=1.0,<2
psutil>=5.9,<8
pyinstaller>=6.0,<7
pywebview>=5.0,<6
